* 'devel' of https://github.com/arangodb/arangodb: (38 commits)
  reformatting, removed potentially undefined behavior
  Doc - setIntroducedIn for log.file-group/mode (#8486)
  Doc - compact, version number (#8521)
  fix compile warning
  The Legendary El Cheapo (#8485)
  fix type of hidden startup option `--rocksdb.recycle-log-file-num` (#8517)
  Fix a race in RemoteExecutor (#8516)
  Revert "Feature/ncc1701 (#8440)"
  Revert "fix a compile warning"
  prevent assertion failures when creating smart graphs with invalid options (#8502)
  fix a compile warning
  use abort as SIGABRT has CORE as default action (#8510)
  Make addfollower jobs less aggressive. (#8490)
  Feature/ncc1701 (#8440)
  attempt to make replication fuzz tests more stable (#8497)
  add RocksDB options `--rocksdb.allow-fallocate` and `--rocksdb.limit-open-files-at-startup` (#8493)
  don't check for version upgrades in enterprise edition when firing up the arangosh (#8495)
  issue 526.3: update analyzer feature to store analyzer definitions in per-vocbase system collections (#8452)
  Doc - Dump program options only if on Enterprise Edition (#8464)
  some very small optimizations (#8484)
  ...

+++ arangod/RocksDBEngine/RocksDBEngine.cpp
@@ -400,6 +400,7 @@ void RocksDBEngine::start() {
   transactionOptions.num_stripes = TRI_numberProcessors();
   transactionOptions.transaction_lock_timeout = opts->_transactionLockTimeout;
 
+  _options.allow_fallocate = opts->_allowFAllocate;
   _options.enable_pipelined_write = opts->_enablePipelinedWrite;
   _options.write_buffer_size = static_cast<size_t>(opts->_writeBufferSize);
   _options.max_write_buffer_number = static_cast<int>(opts->_maxWriteBufferNumber);
@@ -461,7 +462,7 @@ void RocksDBEngine::start() {
   // Maximum number of level-0 files.  We stop writes at this point.
   _options.level0_stop_writes_trigger = static_cast<int>(opts->_level0StopTrigger);
 
-  _options.recycle_log_file_num = static_cast<size_t>(opts->_recycleLogFileNum);
+  _options.recycle_log_file_num = opts->_recycleLogFileNum;
   _options.compaction_readahead_size = static_cast<size_t>(opts->_compactionReadaheadSize);
 
 #ifdef USE_ENTERPRISE
@@ -526,7 +527,14 @@ void RocksDBEngine::start() {
 
   _options.create_if_missing = true;
   _options.create_missing_column_families = true;
-  _options.max_open_files = -1;
+
+  if (opts->_limitOpenFilesAtStartup) {
+    _options.max_open_files = 16;
+    _options.skip_stats_update_on_db_open = true;
+    _options.avoid_flush_during_recovery = true;
+  } else {
+    _options.max_open_files = -1;
+  }
 
   // WAL_ttl_seconds needs to be bigger than the sync interval of the count
   // manager. Should be several times bigger counter_sync_seconds
@@ -706,6 +714,10 @@ void RocksDBEngine::start() {
   // only enable logger after RocksDB start
   if (logger != nullptr) {
     logger->enable();
+  }
+  
+  if (opts->_limitOpenFilesAtStartup) {
+    _db->SetDBOptions({{"max_open_files", "-1"}});
   }
 
   if (_syncInterval > 0) {
@@ -1783,7 +1795,15 @@ void RocksDBEngine::pruneWalFiles() {
     if (deleteFile) {
       LOG_TOPIC(DEBUG, Logger::ENGINES)
           << "deleting RocksDB WAL file '" << (*it).first << "'";
-      auto s = _db->DeleteFile((*it).first);
+      rocksdb::Status s;
+      if (basics::FileUtils::exists(basics::FileUtils::buildFilename(_options.wal_dir, (*it).first))) {
+        // only attempt file deletion if the file actually exists.
+        // otherwise RocksDB may complain about non-existing files and log a big error message
+        s = _db->DeleteFile((*it).first);
+      } else {
+        LOG_TOPIC(DEBUG, Logger::ROCKSDB)
+            << "to-be-deleted RocksDB WAL file '" << (*it).first << "' does not exist. skipping deletion";
+      }
       // apparently there is a case where a file was already deleted
       // but is still in _prunableWalFiles. In this case we get an invalid
       // argument response.
